Implement random state for dice images
