+++ lib/gradient_container.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:first_app/Texts/styled_text.dart';
+
+import 'package:first_app/dice_roller.dart';
 
 const startAlignment = Alignment.topLeft;
 const endAligment = Alignment.bottomRight;
@@ -8,10 +9,6 @@ class GradientContainer extends StatelessWidget {
   const GradientContainer(this.colors, {super.key});
 
   final List<Color> colors;
-
-  void rollDice() {
-    // ...
-  }
 
   @override
   Widget build(context) {
@@ -23,26 +20,8 @@ class GradientContainer extends StatelessWidget {
           end: endAligment,
         ),
       ),
-      child: Center(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Image.asset(
-              'assets/images/dice-2.png',
-              width: 200,
-            ),
-            const SizedBox(
-              height: 20,
-            ),
-            ElevatedButton(
-              onPressed: rollDice,
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.purple,
-              ),
-              child: const StyledText('Roll Dice'),
-            ),
-          ],
-        ),
+      child: const Center(
+        child: DiceRoller(),
       ),
     );
   }
